Add rainforest diagnostics
