+++ homeassistant/components/rainforest_eagle/diagnostics.py
@@ -1,0 +1,23 @@
+"""Diagnostics support for Eagle."""
+from __future__ import annotations
+
+from homeassistant.components.diagnostics import async_redact_data
+from homeassistant.config_entries import ConfigEntry
+from homeassistant.core import HomeAssistant
+
+from .const import CONF_CLOUD_ID, CONF_INSTALL_CODE, DOMAIN
+from .data import EagleDataCoordinator
+
+TO_REDACT = {CONF_CLOUD_ID, CONF_INSTALL_CODE}
+
+
+async def async_get_config_entry_diagnostics(
+    hass: HomeAssistant, config_entry: ConfigEntry
+) -> dict:
+    """Return diagnostics for a config entry."""
+    coordinator: EagleDataCoordinator = hass.data[DOMAIN][config_entry.entry_id]
+
+    return {
+        "config_entry": async_redact_data(config_entry.as_dict(), TO_REDACT),
+        "data": coordinator.data,
+    }
